reconstructing filestructure and adding some features
