+++ commands/fun/__pycache__/slander.cpython-38.pyc
