+++ commands/fun/slander.py
