+++ commands/school/__pycache__/lektieposter.cpython-38.pyc
