+++ commands/school/__pycache__/lektiescanner.cpython-38.pyc
