+++ commands/school/__pycache__/scanschedule.cpython-38.pyc
