+++ commands/school/__pycache__/schedule.cpython-38.pyc
