+++ commands/school/lektiescanner.py
