+++ commands/school/scanschedule.py
