+++ commands/school/schedule.py
