+++ configs/settings.py
@@ -1,0 +1,2 @@
+arg1 = ['lock', 'funnymode']
+arg2 = ['on', 'off']
